update events according to the calendar
